requirements.txt
# 핵심 프레임워크
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# LangGraph & LangChain
langgraph==0.2.50
langchain>=0.3.26
langchain-core>=0.3.66
langchain-openai>=0.3.0

# LLM 연동
openai==1.50.0
ollama==0.4.4

# Confluence API
requests==2.32.3
beautifulsoup4==4.12.3
urllib3==2.2.3

# PDF 생성
xhtml2pdf==0.2.16

# 유틸리티
python-dotenv==1.0.0
pydantic>=2.11.2
anyio>=4.7.0
httpx>=0.27.0

# 테스트
pytest==8.3.0
pytest-asyncio==0.24.0
playwright==1.48.0
pytest-playwright==0.6.2